app/find-it-out_backend/requirements.txt
alembic==1.6.5
bidict==0.21.2
blis==0.7.4
cachelib==0.1.1
catalogue==2.0.4
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.4
Flask-Cors==3.0.10
flask-jwt-extended==4.2.1
Flask-Login==0.4.0
Flask-Migrate==2.0.4
flask-redis==0.4.0
Flask-Session==0.3.2
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5
Flask-WTF==0.14.2
greenlet==1.0.0
gunicorn==19.9.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.4.2
murmurhash==1.0.5
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pathy==0.5.2
Pillow==8.2.0
preshed==3.0.5
progressbar==2.5
psycopg2-binary==2.8.6
pydantic==1.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.1
python-editor==1.0.4
python-engineio==4.1.0
python-magic==0.4.24
python-magic-bin==0.4.14; platform_system == "Windows"
python-resize-image==1.1.19
python-socketio==5.2.1
python-statemachine==0.8.0
redis==3.5.3
regex==2021.4.4
requests==2.10.0
scipy==1.6.3
selenium==3.141.0
shortuuid==1.0.1
six==1.15.0
smart-open==3.0.0
spacy==3.0.6
spacy-legacy==3.0.5
SQLAlchemy==1.4.17
srsly==2.4.1
thinc==8.0.3
tqdm==4.61.0
typer==0.3.2
urllib3==1.26.5
wasabi==0.8.2
Werkzeug==1.0.1
WTForms==2.3.3